+++ src/templates/react-app/api/express/client/api/helpers/restClient.js
@@ -11,7 +11,8 @@ function parseResponseBody(text) {
   }
 }
 
-async function getResponse(response) {
+async function processResponse(responsePromise) {
+  const response = await responsePromise;
   const text = await response.text();
   const body = parseResponseBody(text);
 
@@ -19,56 +20,76 @@ async function getResponse(response) {
     throw new Error(body);
   }
 
-  return { response, body };
+  return body;
 }
 
-function getHeaders() {
-  return {
+function getHeaders(inMemory) {
+  const headers = {
     'Content-Type': 'application/json'
   };
+
+  if (inMemory) {
+    return { ...headers, 'In-Memory': 'true' };
+  }
+
+  return headers;
+}
+
+function getInMemorySetting(method) {
+  const sessionStorageInMemory = sessionStorage.getItem('inMemory');
+  const inMemory = JSON.parse(sessionStorageInMemory ?? '{}');
+  const { all: inMemoryAll, [method]: inMemoryMethod } = inMemory;
+
+  return inMemoryAll || inMemoryMethod;
 }
 
 export async function post(path, body) {
   const url = new URL(path, baseUrl);
+  const inMemory = getInMemorySetting('post');
 
   const response = await fetch(url, {
     method: HttpMethod.Post,
-    headers: getHeaders(),
+    headers: getHeaders(inMemory),
     body: JSON.stringify(body)
   });
 
-  return await getResponse(response);
+  return await processResponse(response);
 }
 
 export async function get(path, id, params) {
-  const urlPath = id ? `${path}/${id}` : path;
+  const urlPath = id !== undefined ? `${path}/${id}` : path;
   const url = new URL(urlPath, baseUrl);
+  const inMemory = getInMemorySetting('get');
 
   const response = await fetch(url, {
     method: HttpMethod.Get,
-    headers: getHeaders()
+    headers: getHeaders(inMemory)
   });
 
-  return await getResponse(response);
+  return await processResponse(response);
 }
 
-export async function put(path, id, body) {
-  const url = new URL(`${path}/${id}`, baseUrl);
+export async function put(path, body) {
+  const url = new URL(path, baseUrl);
+  const inMemory = getInMemorySetting('put');
+
   const response = fetch(url, {
     method: HttpMethod.Put,
-    headers: getHeaders(),
+    headers: getHeaders(inMemory),
     body: JSON.stringify(body)
   });
 
-  return await getResponse(response);
+  return await processResponse(response);
 }
 
 export async function remove(path, id) {
   const url = new URL(`${path}/${id}`, baseUrl);
+  const inMemory = getInMemorySetting('remove');
+
   const response = fetch(url, {
     method: HttpMethod.Delete,
-    headers: getHeaders()
+    headers: getHeaders(inMemory)
   });
 
-  return await getResponse(response);
+  return await processResponse(response);
 }